Edit README.md
Before I edited this, the content said that that "the tool...must be manually deleted" if the output showed pods and replicaset had not been cleaned up. I changed this to say that the pods and replicaset had to be manually deleted. I'm pretty confident that this is correct,  but I wanted to call it out to your attention to make sure it was the right thing to do.